requirements.txt
notebook
pandas
beautifulsoup4
google-cloud-bigquery
numpy
pyarrow
aiohttp